Revert "update discord python starter"
