+++ Dockerfile
@@ -1,39 +1,36 @@
 # https://pythonspeed.com/articles/base-image-python-docker-images/
 # https://testdriven.io/blog/docker-best-practices/
-FROM python:3.11-slim-bullseye
+FROM python:3.10-slim-bullseye
 
-COPY --from=ghcr.io/astral-sh/uv:0.4.9 /uv /bin/uv
+RUN apt-get update && apt-get install -y build-essential
 
-# Set Working directory
+WORKDIR /app
+
+# https://stackoverflow.com/questions/53835198/integrating-python-poetry-with-docker
+ENV PYTHONFAULTHANDLER=1 \
+  PYTHONUNBUFFERED=1 \
+  PYTHONHASHSEED=random \
+  PIP_NO_CACHE_DIR=off \
+  PIP_DISABLE_PIP_VERSION_CHECK=on \
+  PIP_DEFAULT_TIMEOUT=100 \
+  POETRY_VERSION=1.4.1
+
+RUN pip install "poetry==$POETRY_VERSION"
+
+# Copy only requirements to cache them in docker layer
+WORKDIR /app
+COPY poetry.lock pyproject.toml /app/
+
+# Project initialization:
+RUN poetry config virtualenvs.create false \
+  && poetry install --no-root --no-interaction --no-ansi --without dev
+
 WORKDIR /app
 
 RUN addgroup --system app && adduser --system --group app
-RUN chown -R app:app /app
 USER app
 
-# Enable bytecode compilation
-ENV UV_COMPILE_BYTECODE=1
-
-# Copy from the cache instead of linking since it's a mounted volume
-ENV UV_LINK_MODE=copy
-
-# Install the project's dependencies using the lockfile and settings
-RUN --mount=type=cache,target=/root/.cache/uv \
-    --mount=type=bind,source=uv.lock,target=uv.lock \
-    --mount=type=bind,source=pyproject.toml,target=pyproject.toml \
-    uv sync --frozen --no-install-project --no-dev
-
-# Copy only requirements to cache them in docker layer
-COPY uv.lock pyproject.toml /app/
-
-# Sync the project
-RUN --mount=type=cache,target=/root/.cache/uv \
-    uv sync --frozen --no-dev
-
-# Place executables in the environment at the front of the path
-ENV PATH="/app/.venv/bin:$PATH"
-
-COPY --chown=app:app src/ /app/src/
+COPY . .
 
 # https://stackoverflow.com/questions/29663459/python-app-does-not-print-anything-when-running-detached-in-docker
-CMD ["python", "-u", "src/bot.py"]
+CMD ["python", "-u", "-m", "bot.app"]
