+++ pyproject.toml
@@ -1,16 +1,20 @@
-[project]
+[tool.poetry]
 name = "discord-python-starter"
 version = "0.1.0"
 description = "Discord Python Starter Template"
 authors = ["Vineeth Voruganti <13438633+VVoruganti@users.noreply.github.com>"]
 license = "Apache 2.0"
 readme = "README.md"
-requires-python = ">=3.10"
-dependencies = [
-    "py-cord>=2.5.0",
-    "python-dotenv>=1.0.1",
-    "honcho-ai @ git+https://github.com/plastic-labs/honcho-python.git@6f069513e70b3a3600deb09e8eb021337944569f",
-    "openai>=1.42.0",
-    "schedule>=1.2.2",
-    "requests>=2.32.3",
-]
+package-mode = false
+
+[tool.poetry.dependencies]
+python = "^3.10"
+py-cord = "^2.5.0"
+python-dotenv = "^1.0.1"
+honcho-ai = "^0.0.14"
+openai = "^1.42.0"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
